Adds initial outline of idea.
Issues: token sizes of documents:
15011 5228 9911 106976 214597 130525 76082 21183 17941 328130 149508 1710 21662 151974 10617 18174 14400 4623 5478 25103
=> they do not all fit into the context window.
WIP: first create a set of all categories over all documents and then
merge them to a common set.
Then we can do a kind of map-reduce for the common sections to get the
final document with the section.
Or simply keep the common sections.